碰撞系统从float升级为double

+++ src/main/cpp/Collision.h
@@ -8,14 +8,14 @@
 class ChunkManager;
 
 struct AABB {
-    float minX, minY, minZ;
-    float maxX, maxY, maxZ;
+    double minX, minY, minZ;
+    double maxX, maxY, maxZ;
 
     AABB() : minX(0), minY(0), minZ(0), maxX(0), maxY(0), maxZ(0) {}
-    AABB(float minX_, float minY_, float minZ_, float maxX_, float maxY_, float maxZ_)
+    AABB(double minX_, double minY_, double minZ_, double maxX_, double maxY_, double maxZ_)
             : minX(minX_), minY(minY_), minZ(minZ_), maxX(maxX_), maxY(maxY_), maxZ(maxZ_) {}
 
-    AABB offset(float dx, float dy, float dz) const {
+    AABB offset(double dx, double dy, double dz) const {
         return {minX + dx, minY + dy, minZ + dz,
                 maxX + dx, maxY + dy, maxZ + dz};
     }
@@ -26,35 +26,35 @@ struct AABB {
                minZ < other.maxZ && maxZ > other.minZ;
     }
 
-    AABB inflate(float d) const {
+    AABB inflate(double d) const {
         return {minX - d, minY - d, minZ - d,
                 maxX + d, maxY + d, maxZ + d};
     }
 
-    // 新增：获取中心
-    glm::vec3 getCenter() const {
-        return glm::vec3((minX + maxX) * 0.5f, (minY + maxY) * 0.5f, (minZ + maxZ) * 0.5f);
+    // 获取中心
+    glm::dvec3 getCenter() const {
+        return glm::dvec3((minX + maxX) * 0.5, (minY + maxY) * 0.5, (minZ + maxZ) * 0.5);
     }
 
-    // 新增：获取半尺寸
-    glm::vec3 getHalfSize() const {
-        return glm::vec3((maxX - minX) * 0.5f, (maxY - minY) * 0.5f, (maxZ - minZ) * 0.5f);
+    // 获取半尺寸
+    glm::dvec3 getHalfSize() const {
+        return glm::dvec3((maxX - minX) * 0.5, (maxY - minY) * 0.5, (maxZ - minZ) * 0.5);
     }
 
-    // 新增：平移（直接修改）
-    void translate(const glm::vec3& t) {
+    // 平移
+    void translate(const glm::dvec3& t) {
         minX += t.x; minY += t.y; minZ += t.z;
         maxX += t.x; maxY += t.y; maxZ += t.z;
     }
 
-    // 新增：AABB + vec3 偏移（返回新AABB）
-    AABB operator+(const glm::vec3& offset) const {
+    // AABB + dvec3 偏移（返回新AABB）
+    AABB operator+(const glm::dvec3& offset) const {
         return {minX + offset.x, minY + offset.y, minZ + offset.z,
                 maxX + offset.x, maxY + offset.y, maxZ + offset.z};
     }
 
-    // 新增：AABB - vec3（反向偏移）
-    AABB operator-(const glm::vec3& offset) const {
+    // AABB - dvec3（反向偏移）（返回新AABB）
+    AABB operator-(const glm::dvec3& offset) const {
         return {minX - offset.x, minY - offset.y, minZ - offset.z,
                 maxX - offset.x, maxY - offset.y, maxZ - offset.z};
     }
@@ -92,14 +92,14 @@ private:
     std::atomic<float> joystickDY{0.0f};
 
     // 物理状态
-    glm::vec3 position{0.0f, 60.0f, 0.0f};
-    glm::vec3 prevPosition{0.0f, 60.0f, 0.0f};
-    glm::vec3 velocity{0.0f};
-    glm::vec3 stuckSpeedMultiplier{1.0f, 1.0f, 1.0f};
+    glm::dvec3 position{0.0, 60.0, 0.0};
+    glm::dvec3 prevPosition{0.0, 60.0, 0.0};
+    glm::dvec3 velocity{0.0};
+    glm::dvec3 stuckSpeedMultiplier{1.0, 1.0, 1.0};
     bool onGround = false;
     bool horizontalCollision = false;
-    float accumulatedTime = 0.0f;
-    float pitch = 0.0f, yaw = 0.0f;
+    double accumulatedTime = 0.0;
+    double pitch = 0.0, yaw = 0.0;
 
     ChunkManager* chunkManager = nullptr;
     mutable std::mutex mutex;
@@ -110,29 +110,29 @@ private:
     int playerEntityId = -1;
 
     // 常量
-    static constexpr float TICK_DURATION = 1.0f / 20.0f;
-    static constexpr float GRAVITY = 0.08f;
-    static constexpr float JUMP_VELOCITY = 0.42f;
-    static constexpr float PLAYER_WIDTH = 0.6f;
-    static constexpr float PLAYER_HEIGHT = 1.8f;
-    static constexpr float MOVE_SPEED = 0.22f;
-    static constexpr float MOVE_ACCELERATION = 0.15f;
-    static constexpr float SPRINT_MOVE_SPEED = 0.30f;
-    static constexpr float SPRINT_MOVE_ACCELERATION = 0.20f;
-    static constexpr float STEP_HEIGHT = 0.6f;
-    static constexpr float EPSILON = 1e-7f;
+    static constexpr double TICK_DURATION = 1.0 / 20.0;
+    static constexpr double GRAVITY = 0.08;
+    static constexpr double JUMP_VELOCITY = 0.42;
+    static constexpr double PLAYER_WIDTH = 0.6;
+    static constexpr double PLAYER_HEIGHT = 1.8;
+    static constexpr double MOVE_SPEED = 0.22;
+    static constexpr double MOVE_ACCELERATION = 0.15;
+    static constexpr double SPRINT_MOVE_SPEED = 0.30;
+    static constexpr double SPRINT_MOVE_ACCELERATION = 0.20;
+    static constexpr double STEP_HEIGHT = 0.6;
+    static constexpr double EPSILON = 1e-7;
 
     void tick();
     void movePlayer();
     void applyMovement();
-    glm::vec3 collideBoundingBox(const AABB& aabb, const glm::vec3& movement) const;
-    void collideOneAxis(AABB& aabb, glm::vec3& movement, int axis, const std::vector<AABB>& colliders) const;
+    glm::dvec3 collideBoundingBox(const AABB& aabb, const glm::dvec3& movement) const;
+    void collideOneAxis(AABB& aabb, glm::dvec3& movement, int axis, const std::vector<AABB>& colliders) const;
     void checkInsideBlocks();
 
     AABB getPlayerAABB() const;
     // 注意：这里不再重复声明 getBlockAABBs
 
     bool isSwimming() const { return false; } // 暂时留空
-    glm::vec3 getInputVector() const;
-    void applyInputs(float strength);
+    glm::dvec3 getInputVector() const;
+    void applyInputs(double strength);
 };
